% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{create_config}
\alias{create_config}
\title{Create Confif (.json) File}
\usage{
create_config(onto_terms)
}
\arguments{
\item{onto_terms}{a character string of the IRI prefixes to use for mapping (ie, FOODON_00001283, BFO_0000001)}
}
\description{
Create a `config.json` file and output to the temporary directory
}
\keyword{internal}
