% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{create_input}
\alias{create_input}
\title{Create Input}
\usage{
create_input(food_terms)
}
\arguments{
\item{food_terms}{a character string of the food terms you want to map against the FoodOn Ontology}
}
\description{
Create an `input.csv` file andoutput to the temporary directory
}
\keyword{internal}
