% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoodOnMatch.R
\name{FoodOnMatch}
\alias{FoodOnMatch}
\title{Food Ontology Match}
\usage{
FoodOnMatch(food_terms, onto_terms)
}
\arguments{
\item{food_terms}{a vector of food terms to classify}

\item{onto_terms}{a vecor of OBO ontology terms to be included in the mapping (ie, `BFO_0000001`)}
}
\value{
a `tibble` of food term ontoloyy mappings and third-party classifications
}
\description{
Match food terms to the FoodOn Ontology
}
